requirements.txt
google-genai>=1.0.0
tqdm>=4.66.0
Pillow>=10.0.0
python-dotenv>=1.0.0
replicate>=0.25.0
opencv-python>=4.8.0
